WTA #71: Moved InstantProvider to :core:common module.
