+++ home/src/main/java/com/jacob/wakatimeapp/home/HomePageModule.kt
@@ -1,9 +1,6 @@
 package com.jacob.wakatimeapp.home
 
 import com.jacob.wakatimeapp.home.data.network.HomePageAPI
-import com.jacob.wakatimeapp.home.domain.DefaultInstantProvider
-import com.jacob.wakatimeapp.home.domain.InstantProvider
-import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -13,15 +10,9 @@ import retrofit2.Retrofit
 
 @Module
 @InstallIn(SingletonComponent::class)
-@Suppress("UnnecessaryAbstractClass")
-abstract class HomePageModule {
-    @Binds
-    abstract fun provideInstantProvider(impl: DefaultInstantProvider): InstantProvider
-
-    companion object {
-        @Singleton
-        @Provides
-        internal fun provideHomePageService(retrofit: Retrofit) =
-            retrofit.create(HomePageAPI::class.java)
-    }
+object HomePageModule {
+    @Singleton
+    @Provides
+    internal fun provideHomePageService(retrofit: Retrofit) =
+        retrofit.create(HomePageAPI::class.java)
 }
